Completed the Gmail-auto-replier
Using @google-cloud/local-auth and googleapis npm modules,
First the request is authenticated then using the gmail api to get data
from the users inbox. Unread mails that have not been replied to, get sent
an automatic response and then that thread is moved to a label called
'On Vacation'. This process is done for all unread mails that are sent
today. After sending replies to all the available unread mail, it waits
randomly for an interval of 45 - 120 seconds before running again.